Feature/actualizacion ab mde proyectos

+++ mapache-web/src/component/ModuloProyectos/ListadoProyectos.js
@@ -16,29 +16,50 @@ export default class ListadoProyectos extends Component {
         };
     }
 
-    componentDidMount() {
-        axios.get(URL+'proyectos')
-            .then(respuesta => respuesta.data)
-            .then((data) => {
-                this.setState({proyectos : data})
-            });
-    }
-
-    componentDidUpdate(prevProps, prevState, snapshot) {
-        axios.get(URL+'proyectos')
-            .then(respuesta => respuesta.data)
-            .then((data) => {
-                this.setState({proyectos : data})
-            });
-    }
-
-    eliminarProyecto = (proyectoId) => {
-        axios.delete(URL+'proyectos/'+proyectoId)
-            .then(respuesta => {
-                if(respuesta.data != null){
-                    alert("El proyecto fue eliminado correctamente");
+    obtenerProyectos(){
+        (async() => {
+            try {
+                axios.get(URL+'proyectos')
+                    .then(respuesta => respuesta.data)
+                    .then((data) => {
+                        this.setState({proyectos : data})
+                    });
+            } catch (err) {
+                let mensaje = "Error: " + err.response.status;
+                if(err.response.message){
+                    mensaje += ': ' + err.response.message;
                 }
-            });
+                alert(mensaje);
+            }
+        })();
+    }
+
+    componentDidMount() {
+        this.obtenerProyectos();
+    }
+
+    eliminarProyecto = (proyectoId, estado) => {
+        if(estado !== "No iniciado"){
+            alert("Los proyecto ya iniciados no se pueden eliminar");
+            return;
+        }
+        (async() => {
+            try {
+                axios.delete(URL+'proyectos/'+proyectoId)
+                    .then(respuesta => {
+                        if(respuesta.data != null){
+                            alert("El proyecto fue eliminado correctamente");
+                            this.obtenerProyectos();
+                        }
+                    });
+            } catch (err) {
+                let mensaje = "Error: " + err.response.status;
+                if(err.response.message){
+                    mensaje += ': ' + err.response.message;
+                }
+                alert(mensaje);
+            }
+        })();
     }
 
     definirColor(estado){
@@ -97,7 +118,7 @@ export default class ListadoProyectos extends Component {
                                                     Edit
                                                 </Button>
                                                 </Link>
-                                                <Button size="sm" variant="outline-primary" onClick={this.eliminarProyecto.bind(this,proyecto.id)}>
+                                                <Button size="sm" variant="outline-primary" onClick={this.eliminarProyecto.bind(this,proyecto.id, proyecto.estado)}>
                                                     Delete
                                                 </Button>
                                             </ButtonGroup>
